+++ packages/min-dom/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @simple-data-open/min-dom
 
+## 0.0.15
+
+### Patch Changes
+
+- 修复 child 为数组时的渲染, 添加 types datasource 类型
+
 ## 0.0.14
 
 ### Patch Changes